feat (refactor and add spotlight to homepage)
