+++ components/ArtPiecePreview.js
@@ -1,11 +1,11 @@
 import Image from "next/image";
-export default function ArtPiecePreview({ piece }) {
+export default function ArtPiecePreview({ piece, showTitle }) {
   const { artist, name, imageSource } = piece;
   return (
     <li>
       <Image src={imageSource} width={500} height={500} alt={name} />
       <div>
-        <h2>{name}</h2>
+        {showTitle && <h2>{name}</h2>}
         <p>by {artist}</p>
       </div>
     </li>
