+++ pages/index.js
@@ -1,17 +1,5 @@
-import useSWR from "swr";
-import ArtPieces from "@/components/ArtPieces";
+import Spotlight from "@/components/Spotlight";
 
-const url = "https://example-apis.vercel.app/api/art";
-
-export default function HomePage() {
-  const fetcher = (...args) => fetch(...args).then((res) => res.json());
-
-  // fetching data
-  const { data, error, isLoading } = useSWR(url, fetcher);
-
-  // error und loading screen
-  if (error) return <div>Failed to load</div>;
-  if (isLoading) return <div>Loading...</div>;
-
-  return <ArtPieces pieces={data} />;
+export default function HomePage({ pieces }) {
+  return <Spotlight pieces={pieces} showTitle={false} />;
 }
